canary/dns/requirements.txt
dnspython
kafka-python
python-dotenv
